Added ipynb and four images
Added an .ipynb file that reports the area and eccentricities of coins detected in an image, as well as stitching together three images taken of the San Francisco bay to form a panorama. The three photos that form the panorama are called IMG_0.png, IMG_1.png, and IMG_2.png, while the final panorama is called sf_bay_pano.png.